+++ src/libcitygml/geometry.cpp
@@ -1,3 +1,18 @@
+/* -*-c++-*- libcitygml - Copyright (c) 2010 Joachim Pouderoux, BRGM
+*
+* This file is part of libcitygml library
+* http://code.google.com/p/libcitygml
+*
+* libcitygml is free software: you can redistribute it and/or modify
+* it under the terms of the GNU Lesser General Public License as published by
+* the Free Software Foundation, either version 2.1 of the License, or
+* (at your option) any later version.
+*
+* libcitygml is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU Lesser General Public License for more details.
+*/
 ////////////////////////////////////////////////////////////////////////////////
 #include "geometry.hpp"
 #include "polygon.hpp"
@@ -28,17 +43,17 @@ const Envelope& Geometry::getEnvelope( void ) const
 }
 ////////////////////////////////////////////////////////////////////////////////
 // Get the polygons
-unsigned int Geometry::size( void ) const
+size_t Geometry::size( void ) const
 {
     return _polygons.size();
 }
 ////////////////////////////////////////////////////////////////////////////////
-Polygon* Geometry::operator[]( unsigned int i )
+Polygon* Geometry::operator[](size_t i )
 {
     return _polygons[i];
 }
 ////////////////////////////////////////////////////////////////////////////////
-const Polygon* Geometry::operator[]( unsigned int i ) const
+const Polygon* Geometry::operator[]( size_t i ) const
 {
     return _polygons[i];
 }
@@ -97,8 +112,8 @@ bool Geometry::merge( Geometry* g )
 {
     if ( !g || g->_lod != _lod || g->_type != _type ) return false;
 
-    unsigned int pSize = g->_polygons.size();
-    for ( unsigned int i = 0; i < pSize; i++ )
+    size_t pSize = g->_polygons.size();
+    for ( size_t i = 0; i < pSize; i++ )
         _polygons.push_back( g->_polygons[i] );
 
     g->_polygons.clear();
@@ -110,8 +125,8 @@ bool Geometry::merge( Geometry* g )
 ////////////////////////////////////////////////////////////////////////////////
 std::ostream& operator<<( std::ostream& os, const citygml::Geometry& s )
 {
-    unsigned int count = 0;
-    for ( unsigned int i = 0; i < s.size(); i++ )
+    size_t count = 0;
+    for ( size_t i = 0; i < s.size(); i++ )
     {
         os << *s[i];
         count += s[i]->getVertices().size();
