+++ src/libcitygml/geometry.hpp
@@ -1,3 +1,19 @@
+/* -*-c++-*- libcitygml - Copyright (c) 2010 Joachim Pouderoux, BRGM
+*
+* This file is part of libcitygml library
+* http://code.google.com/p/libcitygml
+*
+* libcitygml is free software: you can redistribute it and/or modify
+* it under the terms of the GNU Lesser General Public License as published by
+* the Free Software Foundation, either version 2.1 of the License, or
+* (at your option) any later version.
+*
+* libcitygml is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU Lesser General Public License for more details.
+*/
+////////////////////////////////////////////////////////////////////////////////
 #ifndef __CITYGML_GEOMETRY_HPP__
 #define __CITYGML_GEOMETRY_HPP__
 ////////////////////////////////////////////////////////////////////////////////
@@ -46,9 +62,9 @@ public:
     const Envelope& getEnvelope( void ) const;
 
     // Get the polygons
-    unsigned int size( void ) const;
-    Polygon* operator[]( unsigned int i );
-    const Polygon* operator[]( unsigned int i ) const;
+    size_t size( void ) const;
+    Polygon* operator[]( size_t i );
+    const Polygon* operator[]( size_t i ) const;
 
     const std::vector< Polygon* >& getPolygons() const;
     std::vector< Polygon* >& getPolygons();
